fix: resolve CodeQL code scanning alerts (70 findings)
Production code:
- Fix integer overflow in CostEstimator join cardinality (cast to double
  before multiplication)
- Add using declarations to StringReader in SqlCursorContext,
  SqlSourceTokenizer, and SqlValidator to prevent resource leaks
- Dispose Dialog after Application.Run in SqlQueryScreen
- Remove useless variable assignments in ExecutionPlanBuilder.Statements,
  MergeNode, FetchXmlGenerator, and SqlValidator
- Remove unnecessary null-to-string upcasts in ExecutionPlanBuilder.Join
  and FetchXmlGenerator

Test code:
- Add assertions to empty containers in ScriptExecutionNodeTests,
  TryCatchTests, PpdsDbParameterCollectionTests, and
  PpdsConnectionStringBuilderTests
- Fix integer overflow in PpdsDataReaderTests decimal arithmetic
- Remove useless assignments in CursorNodeTests, RemoteScanNodeTests,
  PpdsDbCommandTests, ExpressionCompilerTests, and
  ExecutionPlanBuilderPhase2Tests
- Remove unnecessary null-to-Object upcasts across join node tests,
  ExpressionCompilerTests, and IsNotDistinctFromTests
- Add null safety to ExpressionCompilerTests FunctionCall casts
- Add using to PpdsDbConnectionTests to ensure dispose on exception

+++ src/PPDS.Query/Planning/CostEstimator.cs
@@ -116,7 +116,7 @@ public sealed class CostEstimator
 
         if (leftCard < 0 || rightCard < 0) return -1;
 
-        return Math.Max(1, (long)(leftCard * rightCard * DefaultJoinSelectivity));
+        return Math.Max(1, (long)((double)leftCard * rightCard * DefaultJoinSelectivity));
     }
 
     private long EstimateParallel(ParallelPartitionNode parallel, CostContext context)
